Module Geoml

module Point : sig ... end

Module for 2d points manipulation

module Circle : sig ... end

Circle manipulation

module Triangle : sig ... end

Triangles manipulation

module Rectangle : sig ... end

This module provides the basic operations over rectangles

module Segment : sig ... end

Segment manipulation

module Line : sig ... end

This module provides basic operation over the linear equation type

module Constraint : sig ... end

Strict and large linear constraints mmanipulation

module Polygon : sig ... end

This module provides basic functions for any concave, convex and convex regular polygon (Does not handle complex polygon and polygon with holes)

module Polyhedron : sig ... end

The module for the polyhedron manipulation. We use the term "polyhedron" to define a set of linear constraint. Each constraint defines a half-plan. A polyhedron is then determined by the intersections of all the half-plans formed by its constaint. It differs from polygons which represent finite enveloppes (For example, convex polygons are a particular cases of polyhedra, which can be infinite).

module Vector : sig ... end