Module Convex.Regular

Module for Regular polygons

type t = private {

center : Point.t;

fst : Point.t;

snd : Point.t;

len : float;

apothem : float;	(*
Distance between the center and the middle of one of the edges
*)

edges : int;

}

val make : Point.t -> Point.t -> int -> t

Create a regular polygon from the center, an arbitrary point and the number of edges

val next_point : ?nth:int -> t -> Point.t

val fold_stop : ( Point.t -> Point.t -> bool ) -> ( int -> 'a -> Point.t -> Point.t -> 'a ) -> 'a -> t -> 'a

val perimeter : t -> float

val area : t -> float

val to_polygon : t -> polygon

val to_randomized_polygon : ?minp:int -> ?prob:float -> t -> polygon

val translate : float -> float -> t -> t

val transform : Geoml__Affine.t -> t -> t

val is_square : t -> bool

val contains : t -> Point.t -> bool

val map : ( Point.t -> Point.t ) -> t -> t