Module Geoml__Curve.Cubic

type t = Geoml.Point.t * Geoml.Point.t * Geoml.Point.t * Geoml.Point.t

This module provides basic operations over cubic bezier curves

val make : 'a -> 'b -> 'c -> 'd -> 'a * 'b * 'c * 'd

Make p0 p1 p2 p3 build a bezier curve starting in p0, and ending in p3, with p1 and p2 as control points

val start : t -> Geoml.Point.t

val ending : t -> Geoml.Point.t

val equation : (Geoml.Point.t * Geoml.Point.t * Geoml.Point.t * Geoml.Point.t) -> float -> Geoml.Point.t

equation t with t belong to 0.;1.

val of_points : Geoml.Point.t list -> (Geoml.Point.t * Geoml.Point.t * Geoml.Point.t * Geoml.Point.t) list

returns the list of curves defined by the given points

val points : (Geoml.Point.t * Geoml.Point.t * Geoml.Point.t * Geoml.Point.t) -> int -> Geoml.Point.t list

points b nb returns a list of nb points uniformly distributed on the bezier curve b.