Module Geoml__Ellipse

Ellipses manipulation

type t = private {

f1 : Geoml.Point.t;

f2 : Geoml.Point.t;

radius : float;

big_axis : float;

small_axis : float;

}

val make_bifocal : Geoml.Point.t -> Geoml.Point.t -> float -> t

val focal1 : t -> Geoml.Point.t

val focal2 : t -> Geoml.Point.t

val center : t -> Geoml.Point.t

val radius : t -> float

val big_axis : t -> float

val small_axis : t -> float

val rotate : t -> Geoml.Point.t -> float -> t

radian rotation. rotate e p f returns the rotated ellipse with p as the rotation center and f a angle in radian

val rotate_angle : t -> Geoml.Point.t -> float -> t

degree rotation. rotate e p f returns the rotated ellipse with p as the rotation center and f a angle in degree

val translate : float -> float -> t -> t

val scale_x : t -> float -> t

val scale_y : t -> float -> t

val contains : t -> Geoml.Point.t -> bool

val area : t -> float