Module Convex.Regular

type t = private {

center : Geoml.Point.t;

fst : Geoml.Point.t;

snd : Geoml.Point.t;

len : float;

apothem : float;
Distance between the center and the middle of one of the edges

edges : int;

}

val make : Geoml.Point.t -> Geoml.Point.t -> int -> t

Create a regular polygon from the center, an arbitrary point and the number of edges

val next_point : ?⁠nth:int -> t -> Geoml.Point.t

val fold_stop : (Geoml.Point.t -> Geoml.Point.t -> bool) -> (int -> 'a -> Geoml.Point.t -> Geoml.Point.t -> 'a) -> 'a -> t -> 'a

val perimeter : t -> float

val area : t -> float

val to_polygon : t -> polygon

val to_randomized_polygon : ?⁠minp:int -> ?⁠prob:float -> t -> polygon

val translate : float -> float -> t -> t

val transform : Geoml__.Affine.t -> t -> t

val is_square : t -> bool

val contains : t -> Geoml.Point.t -> bool

val map : (Geoml.Point.t -> Geoml.Point.t) -> t -> t