Module Geoml__Segment

Segment manipulation

type t = Geoml.Point.t * Geoml.Point.t

val make : Geoml.Point.t -> Geoml.Point.t -> t

module Tbl : Stdlib.Hashtbl.S with type Tbl.key = t

val extr1 : t -> Geoml.Point.t

val extr2 : t -> Geoml.Point.t

val center : t -> Geoml.Point.t

val equation : t -> float -> Geoml.Point.t

val sq_size : t -> float

returns the square size of a segment

val size : t -> float

returns the size of a segment

val scale_x : t -> float -> t

val scale_y : t -> float -> t

val translate : float -> float -> t -> t

val transform : Geoml__.Affine.t -> t -> t

val to_line : t -> Geoml.Line.t

val contains : t -> Geoml.Point.t -> bool

val proj_x : t -> float * float

val proj_y : t -> float * float

val intersects : t -> t -> bool

intersects a b, returns true if a and b intersect. false otherwise

val intersection : t -> t -> Geoml.Point.t option

returns the intersection point of two segments. returns None if they don't intersect

val intersect_line : t -> Geoml.Line.t -> Geoml.Point.t option

returns the intersection point of a segment and a line. returns None if they don't intersect