geoml index

Library geoml

This library exposes the following toplevel modules:

Affine

Circle Circle manipulation

Common

Constraint Strict and large linear constraints mmanipulation

Curve

Ellipse Ellipses manipulation

Geoml

Line This module provides basic operation over the linear equation type

Math

Point Module for 2d points manipulation

Polygon This module provides basic functions for any concave, convex and convex regular polygon (Does not handle complex polygon and polygon with holes)

Polyhedron The module for the polyhedron manipulation. We use the term "polyhedron" to define a set of linear constraint. Each constraint defines a half-plan. A polyhedron is then determined by the intersections of all the half-plans formed by its constaint. It differs from polygons which represent finite enveloppes (For example, convex polygons are a particular cases of polyhedra, which can be infinite).

Polynom

Rectangle This module provides the basic operations over rectangles

Segment Segment manipulation

Triangle Triangles manipulation

Vector